GreetNewMemberService.java
