+++ .github/workflows/deploy-test.yml
@@ -27,6 +27,6 @@ jobs:
         aws_secret_key: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
         application_name: broscorp-community
         environment_name: BroscorpCommunity-test
-        version_label: ${{ GITHUB_SHA }}-test
+        version_label: ${{ github.sha }}-test
         region: eu-west-1
         deployment_package: target/community-tools-1-SNAPSHOT.jar
