+++ .github/workflows/deploy.yml
@@ -19,11 +19,6 @@ jobs:
 
     - name: Generate deployment package
       run: mvn -B package --file pom.xml
-
-    - name: Generate build number
-      uses: einaregilsson/build-number@v2
-      with:
-        token: ${{secrets.github_token}}
       
     - name: Deploy to EB
       uses: einaregilsson/beanstalk-deploy@v4
@@ -32,6 +27,6 @@ jobs:
         aws_secret_key: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
         application_name: broscorp-community
         environment_name: BroscorpCommunity-env
-        version_label: ${{ steps.buildnumber.outputs.build_number }}
+        version_label: ${{ github.sha }}
         region: eu-west-1
         deployment_package: target/community-tools-1-SNAPSHOT.jar
