+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 github.token=${GITHUB_TOKEN}
 github.repository=Broscorp-net/traineeship
 
-slack.webhook="${slack.webhook}"
+slack.webhook="${SLACK_WEBHOOK}"
 slack.token=${SLACK_TOKEN}
 
 # database
